Remove `seccomp` filter
The filter resulted in abnormal process-termination when trying to open an existing database (after daemon restart). My guess as to the reason would be that `badgerhold/v4` uses additional systemcalls, that `v1` did not - or maybe it's a weird side-effect with `systemd-sandboxing`.

Anyways, since `systemd` provides equivalent functionality for services, I don't really see the need for it.